Create takeOutDog.groovy
Items under "trigger when enabled" will occur when the "Take out dog" switch is triggered.
Items under "return trigger" determine what causes the switch to automatically turn off. If you select both a delay and contact then the delay time will be used as a failsafe.
For example, if you set the delay to 10, then after 10 minutes even if the contact is not tripped 2x the system will execute the return procedure.
Items under "trigger when returned" will occur when the trigger is executed.